requirements.txt
asgiref==3.11.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Django==5.2.10
djangorestframework==3.15.2
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==24.2
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.3
Werkzeug==3.1.5